Weather info split to lib folder.
