+++ lib/weather_info/weather_info.cpp
@@ -1,0 +1,16 @@
+#include "weather_info.hpp"
+
+namespace Weather
+{
+void WeatherInfo::clear()
+{
+    name = weather = weather_desc = temperature_c = humidity_perc = counter = "";
+}
+
+String WeatherInfo::get_string()
+{
+    String n = name + ", " + weather + ": " + weather_desc + ". T: " + temperature_c + " C, H: " + humidity_perc +
+               "% ==== [Counter " + counter + "]";
+    return n;
+}
+} // namespace Weather
